Added ascii_common - removed ReadLineResult from ascii_stream and put it in ascii_common, since ReadLineResult is used by both AsciiStreamReader and AsciiStreamReaderAsync
